chore(chat-api): refactor router imports and create routers module
- Updated import paths for appRouter in app.ts, openapi.ts, and server.ts to point to the new routers module.
- Created a new routers.ts file to consolidate router definitions and improve code organization.
